+++ CHANGES.md
@@ -5,6 +5,8 @@ the dispatch function.
 
 Added default for :tool-name option (to `dispatch`), determined from the `babashka.file` system property.
 
+Added :validate directive to command interface, allowing for validations that involve multiple options.
+
 # 0.5 -- 29 Aug 2022
 
 Add public `exit` function as a testable way to call `System/exit`.